requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.12.15; python_version >= '3.9'
aiosignal==1.4.0; python_version >= '3.9'
alabaster==1.0.0; python_version >= '3.10'
annotated-types==0.7.0; python_version >= '3.8'
astroid==3.3.11; python_full_version >= '3.9.0'
attrs==25.3.0; python_version >= '3.8'
azure-core==1.35.0; python_version >= '3.9'
azure-storage-blob==12.26.0; python_version >= '3.8'
babel==2.17.0; python_version >= '3.8'
certifi==2025.8.3; python_version >= '3.7'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.3; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
cryptography==45.0.7; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
datasets==3.6.0; python_full_version >= '3.9.0'
dill==0.3.8; python_version >= '3.8'
docutils==0.21.2; python_version >= '3.9'
filelock==3.19.1; python_version >= '3.9'
frozenlist==1.7.0; python_version >= '3.9'
fsspec[http]==2025.3.0; python_version >= '3.8'
gitdb==4.0.12; python_version >= '3.7'
gitpython==3.1.45; python_version >= '3.7'
hf-xet==1.1.9; python_version >= '3.8'
huggingface-hub==0.34.4; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
imagesize==1.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
iniconfig==2.1.0; python_version >= '3.8'
isodate==0.7.2; python_version >= '3.7'
jinja2==3.1.6; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
mpmath==1.3.0
multidict==6.6.4; python_version >= '3.9'
multiprocess==0.70.16; python_version >= '3.8'
networkx==3.5; python_version >= '3.11'
numpy==2.3.2; python_version >= '3.11'
packaging==25.0; python_version >= '3.8'
pandas==2.3.2; python_version >= '3.9'
platformdirs==4.4.0; python_version >= '3.9'
pluggy==1.6.0; python_version >= '3.9'
propcache==0.3.2; python_version >= '3.9'
protobuf==6.32.0; python_version >= '3.9'
pyarrow==21.0.0; python_version >= '3.9'
pycparser==2.22; python_version >= '3.8'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pygments==2.19.2; python_version >= '3.8'
pytest==8.4.1; python_version >= '3.9'
pytest-asyncio==1.1.0; python_version >= '3.9'
pytest-mock==3.14.1; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.5; python_version >= '3.9'
roman-numerals-py==3.1.0; python_version >= '3.9'
sentry-sdk==2.35.2; python_version >= '3.6'
setuptools==80.9.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smmap==5.0.2; python_version >= '3.7'
snowballstemmer==3.0.1; python_version not in '3.0, 3.1, 3.2'
sphinx==8.2.3; python_version >= '3.11'
sphinx-autoapi==3.6.0; python_version >= '3.9'
sphinxcontrib-applehelp==2.0.0; python_version >= '3.9'
sphinxcontrib-devhelp==2.0.0; python_version >= '3.9'
sphinxcontrib-htmlhelp==2.1.0; python_version >= '3.9'
sphinxcontrib-jsmath==1.0.1; python_version >= '3.5'
sphinxcontrib-qthelp==2.0.0; python_version >= '3.9'
sphinxcontrib-serializinghtml==2.0.0; python_version >= '3.9'
sympy==1.13.1; python_version >= '3.8'
tokenizers==0.22.0; python_version >= '3.9'
torch==2.6.0; python_full_version >= '3.9.0'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
urllib3==2.5.0; python_version >= '3.9'
wandb==0.21.3; python_version >= '3.8'
xxhash==3.5.0; python_version >= '3.7'
yarl==1.20.1; python_version >= '3.9'
